Move file integrity logic to own class
